Added Pytest to the project.
