+++ models/firebase.py
@@ -1,4 +1,4 @@
-from project.db import db
+from db import db
 from typing import Dict, List
