+++ resources/firebase.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource, reqparse
 from typing import Tuple
-from project.models.firebase import FirebaseModel
+from models.firebase import FirebaseModel
 
 
 class FirebaseId(Resource):
@@ -34,7 +34,7 @@ class FirebaseId(Resource):
 
         try:
             fb.save_to_db()
-        except:
+        except :
             return {"message": "An error occurred adding the Firebase item."}, 500
 
         return fb.json(), 200
