+++ setup.cfg
@@ -1,2 +1,9 @@
 [metadata]
 description-file = README.md
+
+[aliases]
+test = pytest
+
+[tool:pytest]
+addopts = -vv
+python_files = tests/*.py
